renamed fuzzy_inference to fuzzy_main
